Getting the entries of the fields in the QuestionsPage done

+++ SubnetMobile/SubnetMobile/Models/IpAddress.cs
@@ -12,22 +12,22 @@ namespace SubnetMobile.Models
         /// <summary>
         /// The first octet or the IP Address.
         /// </summary>
-        public short FirstOctet { get; set; }
+        public int FirstOctet { get; set; }
 
         /// <summary>
         /// The second octet of the IP Address.
         /// </summary>
-        public short SecondOctet { get; set; }
+        public int SecondOctet { get; set; }
 
         /// <summary>
         /// The third octet of the IP Address.
         /// </summary>
-        public short ThirdOctet { get; set; }
+        public int ThirdOctet { get; set; }
 
         /// <summary>
         /// The fourth octet of the IP Address.
         /// </summary>
-        public short FourthOctet { get; set; }
+        public int FourthOctet { get; set; }
 
         public override string ToString()
         {
